add chip display
