+++ src/MainPanel/MainPanel.js
@@ -1,8 +1,6 @@
-import React, { useState, useEffect } from 'react';
 import CardHand from "./Hand.js";
-
-import eventManager from '../Utils/EventManager.js';
-import {calculateHandValue} from "../Game/Card.js"
+import SubmitHandButton from './SubmitHandButton.js';
+import ChipDisplay from './ChipDisplay.js';
 
 import "./styles.css";
 
@@ -10,31 +8,8 @@ export default function MainPanel(){
   return (
     <div className="mainPanel">
       <SubmitHandButton />
+      <ChipDisplay />
       <CardHand />
     </div>
-  )
-}
-
-function SubmitHandButton(){
-  let [buttonText, setButtonText] = useState("+0 Chips")
-
-  useEffect(()=>{
-    const eventHook = eventManager.createHook("updateHandSelection", e => {
-      let value = calculateHandValue(e.hand);
-      setButtonText(`+${value} Chips`)
-    });
-
-    return () => {
-      eventManager.removeHook(eventHook);
-    };
-  });
-
-  function buttonClicked() {
-    eventManager.sendEvent({name: "attemptSubmitHand"});
-  }
-  return (
-    <button className = "submitHandButton" onClick = {buttonClicked}>
-      {buttonText}
-    </button>
   )
 }
